+++ src/pyscomotif/RMSD_calculation.py
@@ -1,6 +1,4 @@
-
 import csv
-from collections import UserDict
 from pathlib import Path
 from typing import Dict, List, Tuple, Callable, Iterator
 
@@ -19,7 +17,7 @@ from pyscomotif.utils import (
     read_compressed_and_pickled_file)
 
 
-def get_n_RMSD_to_compute(PDBs_with_similar_motifs:Dict[nx.Graph, Dict[str, List[nx.Graph]]]) -> int:
+def get_n_RMSD_to_compute(PDBs_with_similar_motifs:Dict[nx.Graph, Dict[str, List[List[Tuple[str,str]]]]]) -> int:
     """
     """
     n_RMSD_to_compute = sum(
@@ -53,16 +51,12 @@ def get_coordinates_of_motif_residues(residue_data: Dict[str,Residue], ordered_r
     return coordinates
 
 def calculate_RMSD_between_two_motifs(
-        similar_motif: nx.Graph, reference_motif_residues_data: Dict[str, Residue], candidate_PDB_residue_data: Dict[str, Residue], RMSD_atoms: str, 
+        target_motif: Tuple[str, ...], query_motif: Tuple[str, ...], reference_motif_residues_data: Dict[str, Residue], candidate_PDB_residue_data: Dict[str, Residue], RMSD_atoms: str, 
     ) -> float:
     """
     """
-    # When performing the sub-graph monomorphism check we saved the residue mapping between the reference motif 
-    # and the similar motif because coordinates have to be paired in advance for quaternion based superimposition to work
-    residue_mapping_dict: Dict[str,str] = similar_motif.residue_mapping_dict
-
-    ordered_reference_residue_IDs: List[str] = [full_residue_ID[:-1] for full_residue_ID in residue_mapping_dict.values()]
-    ordered_target_residue_IDs: List[str] = [full_residue_ID[:-1] for full_residue_ID in residue_mapping_dict.keys()]
+    ordered_reference_residue_IDs: List[str] = [full_residue_ID[:-1] for full_residue_ID in query_motif]
+    ordered_target_residue_IDs: List[str] = [full_residue_ID[:-1] for full_residue_ID in target_motif]
 
     reference_coords = get_coordinates_of_motif_residues(reference_motif_residues_data, ordered_reference_residue_IDs, RMSD_atoms)
     target_coords = get_coordinates_of_motif_residues(candidate_PDB_residue_data, ordered_target_residue_IDs, RMSD_atoms)
@@ -78,19 +72,23 @@ def calculate_RMSD_between_two_motifs(
 
 def calculate_RMSD_values(
         solved_motif_MST: nx.Graph, residue_data_folder_path: Path, compression: str, PDB_ID :str, 
-        list_of_similar_motifs_in_PDB: List[nx.Graph], reference_motif_residues_data: Dict[str, Residue], RMSD_atoms: str
-    ) -> Tuple[nx.Graph, str, UserDict[nx.Graph, float]]:
+        list_of_similar_motifs_in_PDB: List[List[Tuple[str,str]]], reference_motif_residues_data: Dict[str, Residue], RMSD_atoms: str
+    ) -> Tuple[nx.Graph, str, str, Dict[Tuple[Tuple[str, ...], Tuple[str, ...]], float]]:
     """
     """
     candidate_PDB_residue_data = read_compressed_and_pickled_file(residue_data_folder_path / f'{PDB_ID}.{compression}')
     
-    similar_motif_RMSD_map: UserDict[nx.Graph, float] = UserDict()
-    similar_motif_RMSD_map.header_description: str = candidate_PDB_residue_data.header_description # type: ignore
-    for similar_motif in list_of_similar_motifs_in_PDB:
-        RMSD = calculate_RMSD_between_two_motifs(similar_motif, reference_motif_residues_data, candidate_PDB_residue_data, RMSD_atoms)
-        similar_motif_RMSD_map[similar_motif] = RMSD
+    similar_motif_RMSD_map: Dict[Tuple[Tuple[str, ...], Tuple[str, ...]], float] = {}
+    for similar_motif_residue_mapping in list_of_similar_motifs_in_PDB:
+        target_motif = tuple(matched_residues[0] for matched_residues in similar_motif_residue_mapping)
+        query_motif = tuple(matched_residues[1] for matched_residues in similar_motif_residue_mapping)
+        RMSD = calculate_RMSD_between_two_motifs(target_motif, query_motif, reference_motif_residues_data, candidate_PDB_residue_data, RMSD_atoms)
+        
+        similar_motif_RMSD_map[(target_motif, query_motif)] = RMSD
 
-    return (solved_motif_MST, PDB_ID, similar_motif_RMSD_map)
+    header_description: str = candidate_PDB_residue_data.header_description
+
+    return (solved_motif_MST, PDB_ID, header_description, similar_motif_RMSD_map)
 
 def calculate_number_of_mutations(motif_MST: nx.Graph, solved_motif_MST: nx.Graph) -> int:
     """
@@ -104,7 +102,7 @@ def calculate_number_of_mutations(motif_MST: nx.Graph, solved_motif_MST: nx.Grap
     return n_mutations
 
 def calculate_RMSD_between_motif_and_similar_motifs(
-        motif_MST: nx.Graph, PDB_file: Path, PDBs_with_similar_motifs: Dict[nx.Graph, Dict[str, List[nx.Graph]]], 
+        motif_MST: nx.Graph, PDB_file: Path, PDBs_with_similar_motifs: Dict[nx.Graph, Dict[str, List[List[Tuple[str,str]]]]], 
         index_folder_path: Path, RMSD_atoms: str, RMSD_threshold: float, n_cores: int, results_output_path: Path, sort_results: bool
     ) -> None:
     """
@@ -118,10 +116,10 @@ def calculate_RMSD_between_motif_and_similar_motifs(
     )
     
     tqdm_progress_bar = get_tqdm_progress_bar(total=get_n_RMSD_to_compute(PDBs_with_similar_motifs), desc='RMSD calculations')
-    delayed_func: Callable[[nx.Graph, Path, str, str, List[nx.Graph], Dict[str, Residue], str], Tuple[nx.Graph, str, UserDict[nx.Graph, float]]] = delayed(calculate_RMSD_values)
+    delayed_func: Callable[[nx.Graph, Path, str, str, List[List[Tuple[str,str]]], Dict[str, Residue], str], Tuple[nx.Graph, str, str, Dict[Tuple[Tuple[str, ...], Tuple[str, ...]], float]]] = delayed(calculate_RMSD_values)
     with Parallel(n_jobs=n_cores, return_as='generator') as parallel:
         # Each solved motif has a list of PDBs with a similar motif, and each PDB potentially has more than one similar motif
-        results_generator: Iterator[Tuple[nx.Graph, str, UserDict[nx.Graph, float]]] = parallel(
+        results_generator: Iterator[Tuple[nx.Graph, str, str, Dict[Tuple[Tuple[str, ...], Tuple[str, ...]], float]]] = parallel(
             delayed_func(solved_motif_MST, residue_data_folder_path, compression, PDB_ID, list_of_similar_motifs_in_PDB, reference_motif_residues_data, RMSD_atoms)
             for solved_motif_MST, solutions_to_motif in PDBs_with_similar_motifs.items()
                 for PDB_ID, list_of_similar_motifs_in_PDB in solutions_to_motif.items()
@@ -131,18 +129,16 @@ def calculate_RMSD_between_motif_and_similar_motifs(
             csv_writer = csv.writer(file_handle)
             csv_writer.writerow(['matched_motif','similar_motif_found','RMSD','n_mutations','PDB_ID','header_description']) # Column names
 
-            for solved_motif_MST, PDB_ID, similar_motif_RMSD_values_map in results_generator:
-                header_description: str = similar_motif_RMSD_values_map.header_description # type: ignore
-                for similar_motif, RMSD in similar_motif_RMSD_values_map.items(): # Each PDB potentially has more than one similar motif
+            for solved_motif_MST, PDB_ID, header_description, similar_motif_RMSD_values_map in results_generator:
+                for (target_motif, query_motif), RMSD in similar_motif_RMSD_values_map.items(): # Each PDB potentially has more than one similar motif
                     if RMSD > RMSD_threshold:
                         continue
 
-                    residue_mapping_dict: Dict[str,str] = similar_motif.residue_mapping_dict # Mapping between the solved motif's full residue IDs and the full residue IDs of the similar motif found
                     n_mutations = calculate_number_of_mutations(motif_MST, solved_motif_MST)
 
                     row_data: List[str] = [
-                        ' '.join(residue_mapping_dict.values()), # matched_motif
-                        ' '.join(residue_mapping_dict.keys()),   # similar_motif_found
+                        ' '.join(query_motif),
+                        ' '.join(target_motif),
                         str(RMSD),
                         str(n_mutations),
                         PDB_ID,